Added a feature allowing people to use the custom environments @note, @bug, @warning and @todo. These will show up in Bootstrap "alerts" and are colour-coded. I changed the README file accordingly (as well as fixing a few other little problems in it).
Additionally, I reorganized some of my methods so that they are now in a new
module called ford.utils. Previously they were in ford.sourceform, but they were
sufficiently generic that it felt better to put them in a separate module (even
though they aren't used anywhere else yet). This is also where I but the
method to handle the aforementioned new environments.