Adding some initial attempts at parallelising
Prototypes for parallelising:
* HTML generation
* Search node generation
* Graph output

Currently the first two are disabled as they either don't work correctly
or have caused issues (specifically the HTML generation works but uses
incorrect file names meaning all links are broken, the search node
generation seems to work but an issue with the size of the return data
means that this currently hangs for my sample problem).
Currently just the graph output parallelisation is enabled.

Added a new piece of meta data input to control this; the `parallel`
entry determines the maximum number of parallel jobs to use. If this is
<=0 then the standard serial approach will be used otherwise use the
parallel approaches (defaults to 0).