Skip also subdirs of dirs in "exclude_dir".
The previous version was excluding only those directories matching
exactly the entries of exclude_dir, but not the subdirectories.

A special case where the new version can be handy is if one has doc/ as
a subdirectory of src/, since then it suffices excluding src/doc/ to
avoid reading the same file twice.