Added an option mathjax_config for custom setting.
The option mathjax_config can be used to select a javascript file with
additional MathJax settings. As described here:

https://docs.mathjax.org/en/latest/configuration.html#using-plain-javascript

it is possible to define a MathJax object which will then be used in the
initial configuration of MathJax. A typical usage could be defining
custom latex macros as described in

https://docs.mathjax.org/en/latest/tex.html#defining-tex-macros

Hence, given a file my-path/MathJax-my-latex-macros.js containing

window.MathJax = {
  TeX: {
     Macros: {
       RR: "{\\bf R}",
       bold: ["{\\bf #1}",1],
       Abs: ['\\left\\lvert #2 \\right\\rvert_{\\text{#1}}', 2, ""]
     }
   }
}

one can set

mathjax_config: my-path/MathJax-my-latex-macros.js

and use the latex macros \RR, \bold and \Abs anywhere in the ford
documentation.