Introduce a minimum graph depth option and a maximum graph nodes option.
These options can be set via graph_mindepth and graph_maxnodes in the project
meta data. The default for the minimum graph depth is set to 0 to allow for
no generation of the graph at all if the maximum  number of nodes is reached
in the first hop and the default value for the maximum number of nodes is set
to 1000000000 to basically unlimit the node size in a graph.