Introduce the ability to load external modules which are documented with FORD.
If the externalize key is set to true (default is false) FORD saves the information of the modules to a file called modules.pkl. These modules can then be loaded by setting external in the project meta data to the url of this file.

Caution! The current implementation uses pickle for dumping and loading the information. Make sure you can trust the loaded file.