Added a copy_subdir option.
This option allows the user to specify subdirectories that are to be
copied verbatim into the generated documentation directory.

This setting can also be set in the individual index.md within the
page_dir hierarchy.
The option allows a convenient collection of supportive files within
the page_dir hierarchy, for example to provide images.

Example:
project configured with (copy_subdir: media) in the main.md.

Assume the following page_dir structure:

page_dir/
    |- index.md (copy_subdir: images)
                (copy_subdir: plots)
    |- media/
    |- plots/
    |- images/
    |- subtutorial_01/
        |- index.md (copy_subdir: )
        |- media/
    |- subtutorial_02/
        |- index.md (<no option set in file>)
        |- media/

This should produce the a documentation directory as follows:
page/
    |- index.html
    |- plots/
    |- images/
    (media subdirectory not copied, as overwritten in local index.md)
    |- subtutorial_01/
        |- index.html
        (media subdirectory not copied, as overwritten in local index.md)
    |- subtutorial_02/
        |- index.html
        |- media/ (copied due to project settings)