Fix for preprocessors that can't read from stdin
Fixes #389

`gfortran -E` requires a special flag to read from `stdin`, so instead
pass `/dev/null` as a filename. This _should_ be more robust -- at
least, this is how we call the preprocessor elsewhere.

This might fail in other ways for other preprocessors. For instance,
`gfortran -E` emits a warning instead of a preprocessed stream when
passed `/dev/null` like this. It might be necessary to pass an empty
file instead. We also still don't check that flags/macros are passed correctly.